feat(ipfs.ts, cloudflare-images.ts): extract getImageByPath function to cloudflare-images.ts refactor(ipfs.ts): replace inline fetch with getImageByPath function refactor(ipfs.ts): replace body type with IPFSResponseType feat(types.ts): add IPFSResponseType type refactor(ipfs.ts): remove redundant await in fetchIPFS call refactor(ipfs.ts): replace CFIApiResponse with IPFSResponseType
