+++ image/src/utils/cloudflare-images.ts
@@ -94,3 +94,32 @@ export async function urlToCFI({ token, endpoint, imageAccount }: UrlToCFI) {
     imageAccount,
   })
 }
+
+export async function getImageByPath({
+  token,
+  imageAccount,
+  path,
+}: CFImages & { path: string }) {
+  const getImage = await fetch(
+    `https://api.cloudflare.com/client/v4/accounts/${imageAccount}/images/v1/${path}`,
+    {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${token}`,
+      },
+    }
+  )
+  const image = (await getImage.json()) as CFIApiResponse
+
+  if (getImage.ok && image.result) {
+    const variants = image.result.variants
+    const publicURL = variants.find((url) => url.endsWith('/public'))
+
+    if (publicURL) {
+      return publicURL
+    }
+  }
+
+  return ''
+}
